Moved Proxy Vertex
